Day 11 done.

+++ 2022/src/day11/program.h
@@ -18,7 +18,7 @@ public:
   Program();
   void readFile();
   int silver();
-  int gold();
+  long long gold();
 
 private:
   std::vector<std::string> _lines;
